chore: :truck: Move dashboard to its own directory
