+++ frontend/src/routing.tsx
@@ -1,5 +1,5 @@
 import { createBrowserRouter } from 'react-router-dom';
-import App from './dashboard.tsx';
+import App from './dashboard/view.tsx';
 
 export const routing = createBrowserRouter([
   {
